独立出Optimizer IT，建立TPC-H-NO-OPtimizer
