+++ optimizer/src/main/java/cn/edu/tsinghua/iginx/logical/optimizer/rules/ColumnPruningRule.java
@@ -72,6 +72,7 @@ public class ColumnPruningRule extends Rule {
         "ColumnPruningRule",
         "ColumnPruningRule",
         operand(AbstractOperator.class, any()),
+        -1,
         RuleStrategy.ONCE);
   }
